+++ packages/typespec-rust/.scripts/tspcompile.js
@@ -125,6 +125,9 @@ generate('keyvault_secrets', keyvault_secrets, 'test/sdk/keyvault_secrets');
 const blob_storage = pkgRoot + 'test/tsp/Microsoft.BlobStorage/client.tsp';
 generate('blob_storage', blob_storage, 'test/sdk/blob_storage');
 
+const serde_tests = pkgRoot + 'test/tsp/SerdeTests';
+generate('serde_tests', serde_tests, 'test/other/serde_tests');
+
 loopSpec(httpSpecsGroup, httpSpecs)
 loopSpec(azureHttpSpecsGroup, azureHttpSpecs)